Reader atmosphere: extract followErrorMessage into profile-errors
Pulls the action-aware copy helper out of `author-profile-panel.tsx`
and into `profile-errors.ts` so the upcoming followers / following
list views can produce identical follow / unfollow error toasts
without copying the helper. Keeps `not_found` mapping centralized:
"Couldn't follow this account." vs the profile-load shape.
